requirements.txt
altgraph==0.17
beautifulsoup4==4.9.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.0
decorator==4.4.2
future==0.18.2
idna==2.10
lxml==4.5.2
pefile==2019.4.18
pycparser==2.20
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.7
python-telegram-bot==12.8
pywin32-ctypes==0.2.0
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
tornado==6.0.4
urllib3==1.25.10
